+++ libs/miner/include/miner/annealer_miner.hpp
@@ -132,8 +132,9 @@ private:
 
     // DEBUG PRINT OUT SOLUTIONS
 
-    int total_fee = 0;
-    for (auto const &e : generator_.block_fees()) total_fee += e;
+    uint64_t total_fee = 0;
+    for (auto const &e : generator_.block_fees())
+      total_fee += e;
 
     std::cout << "Fee: " << total_fee << std::endl;
     std::cout << "Num Staged: " << generator_.staged().size() << std::endl;
